pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "nuthatch"
authors = [{name = "Benjamin Antieau", email = "antieau@northwestern.edu"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]
dependencies = [
    "python-flint >=0.6.0",
]

[project.urls]
Home = "https://github.com/antieau/nuthatch"
